# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AMX ERP Suite API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface RegisterInput {
  name: string;
  email: string;
  password: string;
  role: string;
}

export interface User {
  id: number;
  name: string;
  email: string;
  role: string;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface DashboardSummary {
  totalRevenue: number;
  totalExpenses: number;
  totalEmployees: number;
  openProjects: number;
  lowStockItems: number;
  pendingInvoices: number;
}

export interface ActivityItem {
  id: number;
  type: string;
  description: string;
  createdAt: string;
}

export interface RevenueDataPoint {
  month: string;
  revenue: number;
  expenses: number;
}

export interface Invoice {
  id: number;
  invoiceNumber: string;
  clientName: string;
  amount: number;
  status: string;
  dueDate: string;
  createdAt: string;
  /** @nullable */
  description?: string | null;
}

export interface InvoiceInput {
  invoiceNumber: string;
  clientName: string;
  amount: number;
  status: string;
  dueDate: string;
  description?: string;
}

export interface InvoiceUpdate {
  clientName?: string;
  amount?: number;
  status?: string;
  dueDate?: string;
  description?: string;
}

export interface Transaction {
  id: number;
  type: string;
  amount: number;
  category: string;
  description: string;
  date: string;
  createdAt: string;
}

export interface TransactionInput {
  type: string;
  amount: number;
  category: string;
  description: string;
  date: string;
}

export interface FinanceSummary {
  totalRevenue: number;
  totalExpenses: number;
  netProfit: number;
  invoicesPaid: number;
  invoicesPending: number;
}

export interface Employee {
  id: number;
  name: string;
  email: string;
  department: string;
  position: string;
  salary: number;
  status: string;
  joinedAt: string;
  createdAt: string;
  /** @nullable */
  phone?: string | null;
}

export interface EmployeeInput {
  name: string;
  email: string;
  department: string;
  position: string;
  salary: number;
  status: string;
  joinedAt: string;
  phone?: string;
}

export interface EmployeeUpdate {
  name?: string;
  email?: string;
  department?: string;
  position?: string;
  salary?: number;
  status?: string;
  phone?: string;
}

export interface Payroll {
  id: number;
  employeeId: number;
  employeeName: string;
  month: number;
  year: number;
  grossSalary: number;
  deductions: number;
  netSalary: number;
  status: string;
  createdAt: string;
}

export interface PayrollInput {
  employeeId: number;
  month: number;
  year: number;
  grossSalary: number;
  deductions: number;
  status: string;
}

export interface HrSummary {
  totalEmployees: number;
  activeEmployees: number;
  departments: number;
  totalPayroll: number;
}

export interface Product {
  id: number;
  name: string;
  sku: string;
  category: string;
  quantity: number;
  price: number;
  reorderLevel: number;
  status: string;
  createdAt: string;
  /** @nullable */
  description?: string | null;
}

export interface ProductInput {
  name: string;
  sku: string;
  category: string;
  quantity: number;
  price: number;
  reorderLevel: number;
  status: string;
  description?: string;
}

export interface ProductUpdate {
  name?: string;
  category?: string;
  quantity?: number;
  price?: number;
  reorderLevel?: number;
  status?: string;
  description?: string;
}

export interface InventorySummary {
  totalProducts: number;
  totalValue: number;
  lowStockCount: number;
  outOfStockCount: number;
}

export interface Project {
  id: number;
  name: string;
  description: string;
  status: string;
  priority: string;
  startDate: string;
  endDate: string;
  budget: number;
  createdAt: string;
  progress?: number;
}

export interface ProjectInput {
  name: string;
  description: string;
  status: string;
  priority: string;
  startDate: string;
  endDate: string;
  budget: number;
  progress?: number;
}

export interface ProjectUpdate {
  name?: string;
  description?: string;
  status?: string;
  priority?: string;
  endDate?: string;
  budget?: number;
  progress?: number;
}

export interface Task {
  id: number;
  projectId: number;
  title: string;
  /** @nullable */
  description?: string | null;
  status: string;
  priority: string;
  /** @nullable */
  assignee?: string | null;
  /** @nullable */
  dueDate?: string | null;
  createdAt: string;
}

export interface TaskInput {
  title: string;
  description?: string;
  status: string;
  priority: string;
  assignee?: string;
  dueDate?: string;
}

export interface TaskUpdate {
  title?: string;
  description?: string;
  status?: string;
  priority?: string;
  assignee?: string;
  dueDate?: string;
}

export interface AnalyticsOverview {
  revenueGrowth: number;
  expenseGrowth: number;
  employeeGrowth: number;
  projectCompletionRate: number;
}

export interface DepartmentSpending {
  department: string;
  amount: number;
}

export interface StatusCount {
  status: string;
  count: number;
}

export interface Notification {
  id: number;
  title: string;
  message: string;
  type: string;
  read: boolean;
  createdAt: string;
}

export interface ForecastDataPoint {
  productName: string;
  currentDemand: number;
  forecastedDemand: number;
  period: string;
}