# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AMX ERP Suite API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /finance/invoices:
    get:
      operationId: listInvoices
      tags: [finance]
      summary: List invoices
      responses:
        "200":
          description: Invoice list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"
    post:
      operationId: createInvoice
      tags: [finance]
      summary: Create invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /finance/invoices/{id}:
    get:
      operationId: getInvoice
      tags: [finance]
      summary: Get invoice
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
    patch:
      operationId: updateInvoice
      tags: [finance]
      summary: Update invoice
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
    delete:
      operationId: deleteInvoice
      tags: [finance]
      summary: Delete invoice
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /finance/transactions:
    get:
      operationId: listTransactions
      tags: [finance]
      summary: List transactions
      responses:
        "200":
          description: Transaction list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
    post:
      operationId: createTransaction
      tags: [finance]
      summary: Create transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
...
  /hr/employees:
    get:
      operationId: listEmployees
      tags: [hr]
      summary: List employees
      responses:
        "200":
          description: Employee list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"
    post:
      operationId: createEmployee
      tags: [hr]
      summary: Create employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /hr/employees/{id}:
    get:
      operationId: getEmployee
      tags: [hr]
      summary: Get employee
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    patch:
      operationId: updateEmployee
      tags: [hr]
      summary: Update employee
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    delete:
      operationId: deleteEmployee
      tags: [hr]
      summary: Delete employee
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /hr/payrolls:
    get:
      operationId: listPayrolls
      tags: [hr]
      summary: List payroll records
      responses:
        "200":
          description: Payroll list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payroll"
    post:
      operationId: createPayroll
      tags: [hr]
      summary: Create payroll entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PayrollInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payroll"
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    RegisterInput:
      type: object
      required: [name, email, password, role]
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, name, email, role, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
        createdAt:
          type: string

    # Dashboard
...
    Invoice:
      type: object
      required: [id, invoiceNumber, clientName, amount, status, dueDate, createdAt]
      properties:
        id:
          type: integer
        invoiceNumber:
          type: string
        clientName:
          type: string
        amount:
          type: number
        status:
          type: string
        dueDate:
          type: string
        createdAt:
          type: string
        description:
          type: ["string", "null"]
    InvoiceInput:
      type: object
      required: [invoiceNumber, clientName, amount, status, dueDate]
      properties:
        invoiceNumber:
          type: string
        clientName:
          type: string
        amount:
          type: number
        status:
          type: string
        dueDate:
          type: string
        description:
          type: string
    InvoiceUpdate:
      type: object
      properties:
        clientName:
          type: string
        amount:
          type: number
        status:
          type: string
        dueDate:
          type: string
        description:
          type: string
    Transaction:
      type: object
      required: [id, type, amount, category, description, date, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
        amount:
          type: number
        category:
          type: string
        description:
          type: string
        date:
          type: string
        createdAt:
          type: string
    TransactionInput:
      type: object
      required: [type, amount, category, description, date]
      properties:
        type:
          type: string
        amount:
          type: number
        category:
          type: string
        description:
          type: string
        date:
          type: string
...
    Employee:
      type: object
      required: [id, name, email, department, position, salary, status, joinedAt, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        department:
          type: string
        position:
          type: string
        salary:
          type: number
        status:
          type: string
        joinedAt:
          type: string
        createdAt:
          type: string
        phone:
          type: ["string", "null"]
    EmployeeInput:
      type: object
      required: [name, email, department, position, salary, status, joinedAt]
      properties:
        name:
          type: string
        email:
          type: string
        department:
          type: string
        position:
          type: string
        salary:
          type: number
        status:
          type: string
        joinedAt:
          type: string
        phone:
          type: string
    EmployeeUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        department:
          type: string
        position:
          type: string
        salary:
          type: number
        status:
          type: string
        phone:
          type: string
    Payroll:
      type: object
      required: [id, employeeId, employeeName, month, year, grossSalary, deductions, netSalary, status, createdAt]
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        employeeName:
          type: string
        month:
          type: integer
        year:
          type: integer
        grossSalary:
          type: number
        deductions:
          type: number
        netSalary:
          type: number
        status:
          type: string
        createdAt:
          type: string
    PayrollInput:
      type: object
      required: [employeeId, month, year, grossSalary, deductions, status]
      properties:
        employeeId:
          type: integer
        month:
          type: integer
        year:
          type: integer
        grossSalary:
          type: number
        deductions:
          type: number
        status:
          type: string
...