# lib/api-zod/src/generated/types/employee.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AMX ERP Suite API
 * OpenAPI spec version: 0.1.0
 */

export interface Employee {
  id: number;
  name: string;
  email: string;
  department: string;
  position: string;
  salary: number;
  status: string;
  joinedAt: string;
  createdAt: string;
  /** @nullable */
  phone?: string | null;
}