# lib/api-zod/src/generated/types/invoiceUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AMX ERP Suite API
 * OpenAPI spec version: 0.1.0
 */

export interface InvoiceUpdate {
  clientName?: string;
  amount?: number;
  status?: string;
  dueDate?: string;
  description?: string;
}