# lib/api-zod/src/generated/types/payrollInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AMX ERP Suite API
 * OpenAPI spec version: 0.1.0
 */

export interface PayrollInput {
  employeeId: number;
  month: number;
  year: number;
  grossSalary: number;
  deductions: number;
  status: string;
}